refactor: move function around
SL-2295
