+++ src/cli/commands/lint.ts
@@ -1,15 +1,19 @@
 import { Command, flags as flagHelpers } from '@oclif/command';
 import { IParserResult } from '@stoplight/types';
 import { getLocationForJsonPath } from '@stoplight/yaml';
+import { writeFile } from 'fs';
 import { resolve } from 'path';
+import { promisify } from 'util';
+import { IRuleResult } from '../..';
+import { json, stylish } from '../../formatters';
 import { readParsable } from '../../fs/reader';
 import { oas2Functions, oas2Rules } from '../../rulesets/oas2';
 import { oas3Functions, oas3Rules } from '../../rulesets/oas3';
 import { readRuleset } from '../../rulesets/reader';
 import { Spectral } from '../../spectral';
 import { IParsedResult, RuleCollection } from '../../types';
-import { formatOutput, writeOutput } from '../utils/output';
 
+const writeFileAsync = promisify(writeFile);
 export default class Lint extends Command {
   public static description = 'lint a JSON/YAML document from a file or URL';
 
@@ -124,4 +128,22 @@ async function lint(name: string, flags: any, command: Lint, customRuleset?: Rul
     process.exitCode = 2;
     throw new Error(ex);
   }
+}
+
+export async function formatOutput(results: IRuleResult[], flags: any): Promise<string> {
+  if (flags.maxResults) {
+    results = results.slice(0, flags.maxResults);
+  }
+  return {
+    json: () => json(results),
+    stylish: () => stylish(results),
+  }[flags.format]();
+}
+
+export async function writeOutput(outputStr: string, flags: any, command: Lint) {
+  if (flags.output) {
+    return writeFileAsync(flags.output, outputStr);
+  }
+
+  command.log(outputStr);
 }
